Use Long for a wider range of User ID's

+++ site/shared/src/main/scala/shared/User.scala
@@ -3,7 +3,7 @@ package shared
 // TODO: is any of this actually used/needed by the client js code?
 
 case class User(
-  id:         Option[Int],
+  id:         Option[Long],
   login:      String,
   name:       String,
   githubId:   String,
